% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawio_engine.R
\name{detect.errors}
\alias{detect.errors}
\title{Detect errors when invoking draw.io}
\usage{
detect.errors(output)
}
\arguments{
\item{output}{The output result from the invocation of \emph{draw.io}, i.e.,
the standard output and error streams (\emph{stdout} and \emph{stderr}) combined.
It should be a vector of characters, where each element is a line.}
}
\value{
Error lines filtered from the \code{output} param, if they exist, or
\code{NULL} otherwise.
}
\description{
This function parses the \emph{draw.io} output to detect errors.
}
\details{
Drawio may encounter errors when invoked, i.e., when exporting a diagram,
for example the diagram source file may be incorrect.
However, they are not visible to the user, because:
\enumerate{
\item \emph{draw.io} is an invoked process that \emph{knitr} does not know about.
\item \emph{draw.io} never returns a non-0 code that would indicate an error.
}

Instead, errors are reported in the output (\emph{stdout} / \emph{stderr}).
These errors need to be detected automatically by \strong{knitrdrawio}, so
that they can be reported to the user, and eventually stop the \emph{knitr}
process if they are blocking.

Additionally, this function is responsible for ignoring known errors that
are not linked to \emph{draw.io} itself (for example, errors related to \emph{dbus}
or \emph{dri3}).
}
