% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drawio.default.path}
\alias{drawio.default.path}
\title{Default path to the Drawio executable}
\usage{
drawio.default.path()
}
\description{
This function returns a (supposedly) sane default for the path to the
\code{drawio} executable, depending on the current OS.
}
\details{
On MacOS (\code{Darwin}), this executable should be located in the
\emph{Applications} directory:
\preformatted{/Applications/draw.io.app/Contents/MacOS/draw.io}

On Linux, the executable is assumed to be located at:
\preformatted{/usr/bin/draw.io}
Note that this path is mostly dependent on the distribution you use,
and the package you installed. On ArchLinux, the \code{/usr/bin} path
should work, but it might be in \code{/bin} in other distributions.

On Windows, it is assumed that the executable is available with
\preformatted{drawio.exe}
}
\note{
This function is only used to return the \emph{default} path, if
no specific path is given by the user. If, on your machine, \code{drawio}
is not available at the default path, or if the OS could not be successfully
identified, please specify your custom location by using the
\code{engine.path} option.

This option can be set in two manners. Either locally, for a single chunk,
by setting it in the code chunk header:
\preformatted{\{drawio mylabel, engine.path = "/path/to/drawio"\}}
Or globally, for all chunks, at the beginning of the RMardown document.
Note that, in this case, you should specify a list of engines (in case you
use other engines, such as \emph{Python} or \emph{Ruby}):
\preformatted{knitr::opts_chunks$set(engine.path = list(
  drawio = "/path/to/drawio",
  # Set here your other engines (if necessary)
  python = "/path/to/python",
))}
}
