% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.os.type}
\alias{get.os.type}
\title{Get (simplified) Operating System type}
\usage{
get.os.type()
}
\description{
This functions returns a simple name for the current OS: "Linux", "Darwin",
or "Windows".
}
