% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_cache.R
\name{hook.cache}
\alias{hook.cache}
\title{Hook function for caching drawio chunks}
\usage{
hook.cache(options)
}
\arguments{
\item{options}{The chunk options}
}
\value{
The chunk options. They are updated with an additional value
(the source file's checksum), indexed by \code{cache.src.md5}, if the
engine is set to \code{drawio} and the \code{cache} option is set to
\code{TRUE}.
}
\description{
This hook is registered with \emph{knitr} to be automatically called when
the user sets the "\code{cache}" chunk option.
If the chunk's engine is "\code{drawio}", this function will compute a MD5
checksum of the specified source file (the "\code{src}" chunk option) and
add it to the chunk options.
As \emph{knitr} uses the chunk options to cache the result of a chunk,
this means that the chunk will not be re-executed as long as the MD5 stays
the same, i.e., the source diagram is not modified.
}
\details{
It should not be called manually by the user, and is registered when the
package is loaded.

If a previous hook was registered for the \code{cache} option, this hook
is first stored so that our hook can call it when executed.
}
