% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_options.R
\name{parse.options}
\alias{parse.options}
\title{Parse Options}
\usage{
parse.options(options)
}
\arguments{
\item{options}{The list of chunk options. Please see in the details for a
list of accepted options.}
}
\value{
cmd The command line that corresponds to the given options.

output The path to the image that will result from the execution
of `\code{cmd}` (including the cache directory, if it was specified).
}
\description{
This function is used to parse the given chunk options and build the
\code{drawio} command that will be executed to export the diagram as an
image.
}
\details{
This function is executed by the drawio engine and should normally not
be used directly by the user.

List of accepted options: (please see drawio's help for more details)
\describe{
  \item{\code{src}}{Path to the source diagram. Mandatory
  argument, must not be null nor empty.}

  \item{\code{format}}{Format of the output image. Either "\code{pdf}",
  "\code{png}", "\code{jpg}", "\code{svg}", "\code{vsdx}", or "\code{xml}".
  If unspecified, a default value is set, depending on the current document
  output: "\code{pdf}" is used for LaTeX, "\code{svg}" for HTML, and
  "\code{png}" for other formats.}

  \item{engine.path}{Path to the \code{draw.io} executable. If unspecified,
  the default value depends on the OS (please see
  \code{\link{drawio.default.path}} for details).}

  \item{engine.opts}{Optional arguments sent as-is to \code{drawio}. Please
  see the \code{drawio} documentation for a list of possible arguments.}

  \item{crop}{Whether to crop the result image or not. (Default: yes)}

  \item{transparent}{Whether to set a transparent background or a blank
  background. Can only be used if `\code{format}` is "\code{pdf}".}

  \item{border}{Width of the border surrounding the diagram. (Default: 0)}

  \item{page.index}{Which page to export from the source diagram, if
  multiple pages are available.}

  \item{page.range}{Similar to `\code{page.index}`, but allowing for
  multiple pages. Can only be used if `\code{format}` is "\code{pdf}".}

  \item{fig.path}{Path to the figure directory. By default, resulting
  images are placed in the current working directory.}
}
}
