% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvfb_wrapper.R
\name{wrap.xvfb}
\alias{wrap.xvfb}
\title{Use a virtual display to run draw.io}
\usage{
wrap.xvfb(command)
}
\arguments{
\item{command}{The list that represents the command that should have been
executed, in a non-headless environment. This list should be the result
of a call to \code{\link{parse.options}}. It must have the following
values: \code{exe}, \code{args}, and \code{output}.}
}
\value{
exe The path to the xvfb-run executable binary.

args The list of command line arguments to be passed to xvfb,
including the path to the draw.io executable, the draw.io arguments,
and electron arguments for headless mode.

output The path to the image that will result from the execution
of `\code{exe} \code{args}` (including the cache directory, if it was
specified). This is exactly the same as the value in the \code{command}
argument.
}
\description{
The draw.io application requires a display to work, which does not exist
in some environments ("headless"). In this case, we can use \code{xvfb} to
emulate a virtual display in place of the physical display.
}
\details{
This method checks that \code{xvfb-run} can be found, and wraps the
command-line call to draw.io as a call to \code{xvfb-run}, with additional
parameters that help draw.io work in headless environments.

The returned value has the same structure as the \code{parse.options} method:
\code{exe} is the path to the executable (in this case, \code{xvfb-run}
instead of draw.io), \code{args} are the arguments, and \code{output} is
left unchanged (path to the output file).
}
