% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_path.R
\name{drawio.default.path}
\alias{drawio.default.path}
\alias{drawio.default.path.linux}
\alias{drawio.default.path.darwin}
\alias{drawio.default.path.windows}
\title{Find the Drawio executable}
\usage{
drawio.default.path()

drawio.default.path.linux()

drawio.default.path.darwin()

drawio.default.path.windows()
}
\description{
This function tries to find the path to a \strong{drawio} executable, assuming
a default installation, in a well-known location, depending on the OS.
}
\section{Functions}{
\itemize{
\item \code{drawio.default.path.linux}: Find default path on Linux

\item \code{drawio.default.path.darwin}: Find default path on Darwin (Mac OS X)

\item \code{drawio.default.path.windows}: Find default path on Windows
}}

\section{Overriding the default path}{


This function is only used to return the \emph{default} path, if
no specific path is given by the user. If, on your machine, \code{drawio}
is not available at the default path, or if the OS could not be successfully
identified, please specify your custom location by using the
\code{engine.path} option.

This option can be set in two manners. Either locally, for a single chunk,
by setting it in the code chunk header:
\preformatted{\{drawio mylabel, engine.path = "/path/to/drawio"\}}
Or globally, for all chunks, at the beginning of the RMardown document.
Note that, in this case, you should specify a list of engines (in case you
use other engines, such as \emph{Python} or \emph{Ruby}):
\preformatted{knitr::opts_chunks$set(engine.path = list(
  drawio = "/path/to/drawio",
  # Set here your other engines (if necessary)
  python = "/path/to/python",
))}
}

\section{Linux}{


On \emph{Linux}, the executable is first searched in the directories specified
in the \code{PATH} environment variable, through the
\code{\link[base]{Sys.which}} function.

If the executable cannot be found in the \code{PATH}, we search
in a few well-known (common) locations:
\itemize{
\item \code{/bin} ;
\item \code{/usr/bin} ;
\item \code{/opt/drawio}.
}
}

\section{MacOS}{


On \emph{Mac OS X} (\code{Darwin}), the executable is first searched in
the directories specified in the \code{PATH} environment variable,
through the \code{\link[base]{Sys.which}} function.

If the executable cannot be found in the \code{PATH}, we search
in a few well-known (common) locations:
\itemize{
\item \code{/Applications} ;
\item \code{~/Applications} ;
\item \code{~/bin} ;
\item \code{/bin} ;
\item \code{/usr/bin} ;
\item \code{/usr/local/bin} ;
\item \code{/opt/drawio}.
}
}

\section{Windows}{


On \emph{Windows}, the executable is first searched in the directories specified
in the \code{PATH} environment variable, through the
\code{\link[base]{Sys.which}} function.
}

